backend/pyproject.toml
[tool.poetry]
name = "financial-backend"
version = "1.0.0"
description = "Financial Backend API for Portfolios"
authors = ["Dev Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pandas = "^2.2.0"
greenlet = "^3.3.0"
psycopg2-binary = "^2.9.9"
authlib = "^1.3.0"
httpx = "^0.27.0"
pdfplumber = "^0.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"